api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
slowapi>=0.1.9
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
# Install the skycli package from repo root
.
